Split battle bridge and persistence out of TitanGame into game/
game.ts mixed movement/turn orchestration with two self-contained
concerns: the thin pass-through layer to Battle (getBattle*, mBattle*,
doBattle* actions) and save/load logic (persist, mRehydrate, doRestore).
Both extract cleanly since neither touches the split/move/muster
orchestration that makes the rest of the class resist a further split.

The store reflects over TitanGame.prototype to wire up Vuex
getters/mutations/actions, so the extracted methods are applied via
Object.assign(TitanGame.prototype, gameBattle, gamePersistence) rather
than declared in the class body, with TypeScript interface merging
keeping them type-checked on TitanGame itself. This mirrors the pattern
already used for BattleCreature in
src/models/battle/combatant.ts.

No behavior change; code moved, not rewritten. Verified
Object.getOwnPropertyNames(TitanGame.prototype) is identical before and
after via a throwaway comparison.
